refactor: extract resolveNodeConnectionParams from node task action
Move out the resolution code for network connections (the mapping code
taking the cli arguments and generating the EDR config settings).

This is a first step to bring this logic under test.
